Spawning entities working. Moving onto collisions

+++ src/entity.c
@@ -4,13 +4,13 @@
 #include <SDL_image.h>
 #include "obj.h"
 #include "boundingbox.h"
-
 #include "entity.h"
+#include "simple_logger.h"
 
 Entity entList[MAX_ENT];
 int ent_count;
 
-Entity* Spawn_Ent()
+Entity* ent_Spawn()
 {
 	int i;
 
@@ -24,12 +24,34 @@ Entity* Spawn_Ent()
 			return &entList[i];
 		}
 		
-		fprintf ( stderr, "Max entities reached. Please kill some entities before making more." );
+		if(i == MAX_ENT)
+			fprintf ( stderr, "Max entities reached. Please kill some entities before making more." );
 	}
 	return NULL;
 }
 
-void entDraw(Entity *ent)
+Entity* ent_New(const char *name, Vec3D position, BoundingVolume bv)
+{
+	Entity *ent;
+	ent = ent_Spawn();
+	if(!ent) return;
+	
+	if(bv.selection == 2)
+		ent->objModel = obj_load("models/cube.obj");
+	if(bv.selection == 1)
+		ent->objModel = obj_load("models/sphere.obj");
+	ent->body.position = position;
+	ent->body.rotation = vec3d(0,0,0);
+	ent->body.scale = vec3d(0.5,0.5,0.5);
+	setBoundingVolume(ent->body.ent_BB, bv);
+	slog("%s 's bounding volume set.", name);
+	ent->texture = NULL;
+	ent->colour = vec4d(1,1,1,0.5);
+
+	return ent;
+}
+
+void ent_Draw(Entity *ent)
 {
 	if (!ent) return;
 
@@ -41,10 +63,22 @@ void entDraw(Entity *ent)
 		ent->colour,
 		ent->texture
 	);
-
 }
 
-void KillEnt(Entity* ent)
+void ent_DrawAll()
+{
+    int i;
+    for (i = 0;i < 1048/*ENT_MAX*/;i++)
+    {
+        if (entList[i].inuse)
+        {
+            ent_Draw(&entList[i]);
+			//slog("Entity drawn!");
+        }
+    }
+}
+
+void ent_Kill(Entity* ent)
 {
 	ent_count--;
 	ent->inuse = 0;
